+++ dp/_dp_array.py
@@ -1,6 +1,6 @@
 """This file provides the DPArray class."""
+from typing import Iterable
 import numpy as np
-
 from dp._logger import Logger, Op
 
 
@@ -97,7 +97,7 @@ class DPArray:
             # Convert 1d slice to nd slice.
             nd_slice = (nd_slice,)
         if not isinstance(nd_slice, (list, tuple)):
-            raise ValueError(f"'nd_slice' has type {type(nd_slice)}, must be"
+            raise ValueError(f"'nd_slice' has type {type(nd_slice)}, must be "
                              f"a slice object, a list/tuple of slice objects,"
                              f"or a list/tuple of integers.")
 
@@ -198,6 +198,113 @@ class DPArray:
         if isinstance(other, DPArray):
             return self.arr != other.arr
         return self.arr != other
+
+    def _max_min(self, cmp, refs, preprocessing, const):
+        """
+        Args:
+            cmp (callable): Use x > y for max and x < y for min
+            idx: The index to assign the calculated value to
+            refs (iterable of indices): Indices to retreive values
+                from to use in the max/min function. Must be an
+                iterable even if the iterable is a singleton
+            preprocessing (callable or iterable of callables): 
+                If callable preprocessing will be applied to each
+                ref value before applying the max/min function. 
+                If iterable of callables, then it is requried the
+                len(refs) = len(preprocessing). preprocessing[i]
+                will be applied to refs[i] before applying the max function.
+            const (float): A constant value to use in the min/max operation
+                
+        Returns:
+            The max/min value of the references after applying preprocessing  
+        """
+        # Error handling
+        if not isinstance(refs, Iterable) or len(refs) == 0:
+            raise ValueError(
+                "Expecting reference to be Iterable of length " + \
+                "at least one."
+            )
+        if not callable(preprocessing) and len(preprocessing) != len(refs):
+            raise ValueError(
+                "Expected refs and preprocessing of same length or single " + \
+                "preprocessing callable."
+            )
+
+        # Make iterable to iterate over
+        if callable(preprocessing):
+            itr = [(ref, preprocessing) for ref in refs]
+        else:
+            itr = zip(refs, preprocessing)
+
+        # Find max/min value and corresponding idx
+        best_idx = None
+        best_val = const
+        for ref, func in itr:
+            val = func(self[ref])
+
+            if isinstance(val, np.ndarray):
+                slice_indices = self._nd_slice_to_indices(ref)
+                val = val.flatten()
+                slice_max_idx = val.argmax()
+                val = val[slice_max_idx]
+                ref = slice_indices[slice_max_idx]
+
+            if best_val is None or cmp(val, best_val):
+                best_idx = ref
+                best_val = val
+
+        # Highlight and write value
+        if best_idx is not None:
+            self.logger.append(self._array_name, Op.HIGHLIGHT, best_idx)
+        return best_val
+
+    def max(self, refs, preprocessing=(lambda x: x), const=None):
+        """
+        Args:
+            idx: The index to assign the calculated value to
+            refs (iterable of indices): Indicies to retreive
+                values from to use in the max function
+            preprocessing (callable or iterable of callables): 
+                If callable preprocessing will be applied to
+                each refs value before applying the max function. 
+                If iterable of callables, then it is requried the
+                len(refs) = len(preprocessing). preprocessing[i]
+                will be applied to refs[i] before applying the max
+                function.
+            const (float): A constant value to use in the min/max
+                operation
+                
+        Returns:
+            The maximum value after applying preprocessing to refs             
+        """
+        return self._max_min(cmp=lambda x, y: x > y,
+                             refs=refs,
+                             preprocessing=preprocessing,
+                             const=const)
+
+    def min(self, refs, preprocessing=(lambda x: x), const=None):
+        """
+        Args:
+            idx: The index to assign the calculated value to
+            refs (iterable of indices): Indicies to retreive
+                values from to use in the min function
+            preprocessing (callable or iterable of callables): 
+                If callable preprocessing will be applied to
+                each refs value before applying the min function. 
+                If iterable of callables, then it is requried the
+                len(refs) = len(preprocessing). preprocessing[i]
+                will be applied to refs[i] before applying the min
+                function.
+            const (float): A constant value to use in the min/max
+                operation
+                
+        Returns:
+            The minimum value after applying preprocessing to refs       
+        """
+        return self._max_min(cmp=lambda x, y: x < y,
+                             refs=refs,
+                             preprocessing=preprocessing,
+                             const=const)
 
     @property
     def arr(self):